+++ src/Facade/FacadeTrait.php
@@ -19,6 +19,7 @@ use Eloquent\Phony\Spy\SpyVerifier;
 use Eloquent\Phony\Stub\StubVerifier;
 use InvalidArgumentException;
 use ReflectionClass;
+use ReflectionType;
 use Throwable;
 
 /**
@@ -280,6 +281,18 @@ trait FacadeTrait
     ): WildcardMatcher {
         return static::driver()->matcherFactory
             ->wildcard($value, $minimumArguments, $maximumArguments);
+    }
+
+    /**
+     * Get an "empty" value for the supplied type.
+     *
+     * @param ReflectionType $type The type.
+     *
+     * @return mixed An "empty" value of the supplied type.
+     */
+    public static function emptyValue(ReflectionType $type)
+    {
+        return static::driver()->emptyValueFactory->fromType($type);
     }
 
     /**