+++ src/functions.php
@@ -20,6 +20,7 @@ use Eloquent\Phony\Spy\SpyVerifier;
 use Eloquent\Phony\Stub\StubVerifier;
 use InvalidArgumentException;
 use ReflectionClass;
+use ReflectionType;
 use Throwable;
 
 /**
@@ -274,6 +275,18 @@ function wildcard(
 ): WildcardMatcher {
     return FacadeDriver::instance()->matcherFactory
         ->wildcard($value, $minimumArguments, $maximumArguments);
+}
+
+/**
+ * Get an "empty" value for the supplied type.
+ *
+ * @param ReflectionType $type The type.
+ *
+ * @return mixed An "empty" value of the supplied type.
+ */
+function emptyValue(ReflectionType $type)
+{
+    return FacadeDriver::instance()->emptyValueFactory->fromType($type);
 }
 
 /**