+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 - **[BC BREAK]** PHP 5 is no longer supported.
+- **[NEW]** Implemented `emptyValue()`.
 - **[IMPROVED]** Support for PHP 7.2 features, including the `object` typehint.
 - **[IMPROVED]** Reduced the amount of output generated when mocks, stubs, and
   spies are encountered by `var_dump()`.